+++ R/stmtPlanner.R
@@ -174,6 +174,7 @@ quote_mutate <- function(...) {
 #'
 mutate_seb <- function(d, blocks,
                        env = parent.frame()) {
+  force(env)
   if(!(is.data.frame(d) || dplyr::is.tbl(d))) {
     stop("seplyr::mutate_seb first argument must be a data.frame or tbl")
   }