+++ R/transmute_se.R
@@ -26,6 +26,7 @@ transmute_se <- function(.data, transmuteTerms,  env=parent.frame()) {
   if(!(is.data.frame(.data) || dplyr::is.tbl(.data))) {
     stop("seplyr::transmute_se first argument must be a data.frame or tbl")
   }
+  force(env)
   # convert char vector into spliceable vector
   # from: https://github.com/tidyverse/rlang/issues/116
   # updated: https://github.com/WinVector/seplyr/issues/3